refactoring DropoutNd function to ml4gw/nn/dropout.py
