+++ ml4gw/nn/__init__.py
@@ -5,4 +5,4 @@ to get started, rather than defining your own
 architecture from the start.
 """
 
-from . import autoencoder, norm, resnet, ssm, streaming
+from . import autoencoder, dropout, norm, resnet, ssm, streaming
